TEMP PATCH.
This enables support for asserts in initializer lists. However, it also
hacks dependency overrides on analyzer and front_end since the needed
versions of those packages are not published yet.

If you use this branch, you'll need to hack the pubspec to point to
your own local copies of those packages from the Dart SDK.